+++ src/features/profiles/utils/profiles-actions.functions.ts
@@ -2,6 +2,7 @@ import { createServerFn } from '@tanstack/react-start';
 import { zodValidator } from '@tanstack/zod-adapter';
 import { personalInfoSchema, techStackSchema, socialLinkSchema, UserOutput } from './schemas';
 import {
+  getProfilePresentationInternal,
   getProfileInternal,
   updatePersonalInfoInternal,
   addTechStackInternal,
@@ -13,7 +14,14 @@ import {
 import { z } from 'zod';
 
 /**
- * Server function to fetch the current user's profile.
+ * Server function to fecth to the front page.
+ */
+export const getProfilePresentationAction = createServerFn({ method: 'GET' }).handler(async (): Promise<UserOutput> => {
+  return await getProfilePresentationInternal();
+});
+
+/**
+ * Server function to fetch the current user's profile to admin page.
  */
 export const getProfileAction = createServerFn({ method: 'GET' }).handler(async (): Promise<UserOutput> => {
   return await getProfileInternal();